refcator: 카테고리 컴포넌트 이름 변경 및 prop 구조 간소화
