+++ frontend/apps/schedule/components/CategorySelect/Category.css.ts
